Add support for bind mounts
This commit introduces support for bind mounting directories
into a LXC. For this, a new config option mounts and a new
command setup_ssh is introduced that will be invoed in the
create command.

The mounts config option is an array of directory pairs with
the source directory in the host filesystem and the target
directory in the container filesystem. The lxc.conf option
lxc.mount.entry is used for to mount these when starting the
container.